restructure abstraction
