+++ scripts/arbin
@@ -1,0 +1,14 @@
+# coding: utf-8
+# Copyright 2025 National Technology & Engineering Solutions of Sandia, LLC (NTESS). Under the terms of Contract DE-NA0003525 with NTESS, the U.S. Government retains certain rights in this software.
+
+import pandas as pd
+
+class arbin(abstractFileType):
+
+    def file_to_df(path, name, **kwargs):
+        #reads yaml function name to chose pandas function for file reading and reads
+        read_func = getattr(pd, name)
+        if name=='read_excel': #is there a better way to choose this if needed?
+            return read_func(path)(**kwargs)
+        else:
+            return read_func(path)
